# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ChartData": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Id",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ChartSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "unit_code": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Unit Code",
            "type": "string"
          }
        },
        "required": ["id", "unit_code", "series"],
        "title": "ChartData",
        "type": "object"
      },
      "ChartPoint": {
        "additionalProperties": false,
        "properties": {
          "value": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "x": {
            "maxLength": 100,
            "minLength": 1,
            "title": "X",
            "type": "string"
          }
        },
        "required": ["x", "value"],
        "title": "ChartPoint",
        "type": "object"
      },
      "ChartSeries": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Id",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ChartPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": ["id", "points"],
        "title": "ChartSeries",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalizedElementText": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "series_labels": {
            "patternProperties": {
              "^[a-z][a-z0-9_.-]{0,99}$": {
                "type": "string"
              }
            },
            "title": "Series Labels",
            "type": "object"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "unit_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Label"
          }
        },
        "required": ["title"],
        "title": "LocalizedElementText",
        "type": "object"
      },
...
      "MarketPolicyUpdate": {
        "properties": {
          "contract_reference": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Contract Reference",
            "type": "string"
          },
          "is_visible": {
            "title": "Is Visible",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": ["is_visible", "contract_reference", "reason"],
        "title": "MarketPolicyUpdate",
        "type": "object"
      },
      "MarketResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "is_visible": {
            "title": "Is Visible",
            "type": "boolean"
          },
          "name_en": {
            "title": "Name En",
            "type": "string"
          },
          "name_zh_hans": {
            "title": "Name Zh Hans",
            "type": "string"
          },
          "name_zh_hant": {
            "title": "Name Zh Hant",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name_en",
          "name_zh_hant",
          "name_zh_hans",
          "is_visible"
        ],
        "title": "MarketResponse",
        "type": "object"
      },
...
      "MetricValue": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Id",
            "type": "string"
          },
          "unit_code": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Unit Code",
            "type": "string"
          },
          "value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": ["id", "value", "unit_code"],
        "title": "MetricValue",
        "type": "object"
      },
...
      "PresentationContract": {
        "additionalProperties": false,
        "properties": {
          "labels": {
            "patternProperties": {
              "^[a-z][a-z0-9_.-]{0,99}$": {
                "$ref": "#/components/schemas/LocalizedElementText"
              }
            },
            "title": "Labels",
            "type": "object"
          },
          "locale": {
            "enum": ["zh-TW", "zh-CN", "en"],
            "title": "Locale",
            "type": "string"
          },
          "schema_version": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Schema Version",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["schema_version", "locale", "title", "labels"],
        "title": "PresentationContract",
        "type": "object"
      },
...
      "PublicationContent": {
        "additionalProperties": false,
        "description": "Locale-neutral application-owned values.\n\nDecimal is intentionally retained in memory and serialized as a JSON string\nby Pydantic. This avoids locale-specific formatting and binary float loss.",
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "charts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChartData"
            },
            "title": "Charts",
            "type": "array"
          },
          "market_code": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Market Code",
            "type": "string"
          },
          "metrics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MetricValue"
            },
            "title": "Metrics",
            "type": "array"
          },
          "schema_version": {
            "pattern": "^[a-z][a-z0-9_.-]{0,99}$",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": ["schema_version", "market_code", "as_of"],
        "title": "PublicationContent",
        "type": "object"
      },
...
      "ReportPublicationResponse": {
        "properties": {
          "content": {
            "$ref": "#/components/schemas/PublicationContent"
          },
          "edition_date": {
            "format": "date",
            "title": "Edition Date",
            "type": "string"
          },
          "locale": {
            "enum": ["zh-TW", "zh-CN", "en"],
            "title": "Locale",
            "type": "string"
          },
          "market_code": {
            "title": "Market Code",
            "type": "string"
          },
          "presentation": {
            "$ref": "#/components/schemas/PresentationContract"
          },
          "publication_id": {
            "format": "uuid",
            "title": "Publication Id",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "report_key": {
            "title": "Report Key",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "source_as_of": {
            "format": "date",
            "title": "Source As Of",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "stale_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Reason"
          }
        },
        "required": [
          "publication_id",
          "report_key",
          "market_code",
          "edition_date",
          "revision",
          "source_as_of",
          "published_at",
          "stale",
          "stale_reason",
          "content",
          "presentation",
          "locale"
        ],
        "title": "ReportPublicationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Daily Insights API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/organizations/{organization_id}/markets": {
      "get": {
        "operationId": "list_organization_markets_api_admin_organizations__organization_id__markets_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketResponse"
                  },
                  "title": "Response List Organization Markets Api Admin Organizations  Organization Id  Markets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Organization Markets",
        "tags": ["administration"]
      }
    },
    "/api/admin/organizations/{organization_id}/markets/{market_code}": {
      "put": {
        "operationId": "set_organization_market_api_admin_organizations__organization_id__markets__market_code__put",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "market_code",
            "required": true,
            "schema": {
              "title": "Market Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketPolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Organization Market",
        "tags": ["administration"]
      }
    },
...
    "/api/reports": {
      "get": {
        "operationId": "list_latest_reports_api_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "zh-TW",
              "enum": ["zh-TW", "zh-CN", "en"],
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "report_key",
            "required": false,
            "schema": {
              "default": "daily-market",
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[a-z][a-z0-9_.-]*$",
              "title": "Report Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportPublicationResponse"
                  },
                  "title": "Response List Latest Reports Api Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Latest Reports",
        "tags": ["reports"]
      }
    },
    "/api/reports/{market_code}/latest": {
      "get": {
        "operationId": "get_latest_report_api_reports__market_code__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "market_code",
            "required": true,
            "schema": {
              "title": "Market Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "zh-TW",
              "enum": ["zh-TW", "zh-CN", "en"],
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "report_key",
            "required": false,
            "schema": {
              "default": "daily-market",
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[a-z][a-z0-9_.-]*$",
              "title": "Report Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPublicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Report",
        "tags": ["reports"]
      }
    }
  }
}